requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0 